Finally do away with ingot_packet.rs
