+++ lib/opte-test-utils/src/lib.rs
@@ -30,8 +30,6 @@ pub use opte::engine::geneve::GENEVE_OPT_CLASS_OXIDE;
 pub use opte::engine::geneve::GENEVE_PORT;
 pub use opte::engine::headers::IpAddr;
 pub use opte::engine::headers::IpCidr;
-pub use opte::engine::ingot_packet::MblkLiteParsed;
-pub use opte::engine::ingot_packet::Packet;
 pub use opte::engine::ip::v4::Ipv4;
 pub use opte::engine::ip::v4::Ipv4Addr;
 pub use opte::engine::ip::v4::Protocol;
@@ -39,6 +37,8 @@ pub use opte::engine::ip::v6::Ipv6;
 pub use opte::engine::ip::v6::Ipv6Addr;
 pub use opte::engine::ip::L3Repr;
 pub use opte::engine::layer::DenyReason;
+pub use opte::engine::packet::MblkLiteParsed;
+pub use opte::engine::packet::Packet;
 pub use opte::engine::packet::ParseError;
 pub use opte::engine::port::meta::ActionMeta;
 pub use opte::engine::port::DropReason;
