+++ lib/opte/src/engine/icmp/v6.rs
@@ -9,9 +9,9 @@
 use super::*;
 use crate::ddi::mblk::MsgBlk;
 use crate::engine::ether::Ethernet;
-use crate::engine::ingot_packet::MblkPacketData;
 use crate::engine::ip::v6::Ipv6;
 use crate::engine::ip::v6::Ipv6Ref;
+use crate::engine::packet::MblkPacketData;
 use crate::engine::predicate::Ipv6AddrMatch;
 use alloc::string::String;
 use ingot::ethernet::Ethertype;
