+++ lib/opte/src/engine/parse.rs
@@ -24,7 +24,6 @@ use super::headers::ValidEncapMeta;
 use super::icmp::IcmpEchoMut;
 use super::icmp::QueryEcho;
 use super::icmp::ValidIcmpEcho;
-use super::ingot_packet::OpteMeta;
 use super::ip::v4::Ipv4Mut;
 use super::ip::v4::Ipv4Ref;
 use super::ip::v6::Ipv6Mut;
@@ -35,6 +34,7 @@ use super::ip::L3;
 use super::packet::AddrPair;
 use super::packet::InnerFlowId;
 use super::packet::MismatchError;
+use super::packet::OpteMeta;
 use super::packet::ParseError;
 use super::packet::FLOW_ID_DEFAULT;
 use super::rule::CompiledTransform;
