+++ lib/opte/src/engine/port.rs
@@ -18,11 +18,6 @@ use super::headers::EncapPush;
 use super::headers::HeaderAction;
 use super::headers::IpPush;
 use super::headers::UlpHeaderAction;
-use super::ingot_packet::FullParsed;
-use super::ingot_packet::LiteParsed;
-use super::ingot_packet::MblkFullParsed;
-use super::ingot_packet::MblkPacketData;
-use super::ingot_packet::Packet;
 use super::ioctl;
 use super::ioctl::TcpFlowEntryDump;
 use super::ioctl::TcpFlowStateDump;
@@ -38,7 +33,12 @@ use super::layer::LayerStatsSnap;
 use super::layer::RuleId;
 use super::packet::BodyTransform;
 use super::packet::BodyTransformError;
+use super::packet::FullParsed;
 use super::packet::InnerFlowId;
+use super::packet::LiteParsed;
+use super::packet::MblkFullParsed;
+use super::packet::MblkPacketData;
+use super::packet::Packet;
 use super::packet::FLOW_ID_DEFAULT;
 use super::rule::Action;
 use super::rule::CompiledTransform;
@@ -67,8 +67,8 @@ use crate::ddi::sync::KMutex;
 use crate::ddi::sync::KMutexType;
 use crate::ddi::time::Moment;
 use crate::engine::flow_table::ExpiryPolicy;
-use crate::engine::ingot_packet::EmitSpec;
-use crate::engine::ingot_packet::PushSpec;
+use crate::engine::packet::EmitSpec;
+use crate::engine::packet::PushSpec;
 use crate::engine::rule::CompiledEncap;
 use crate::ExecCtx;
 use alloc::boxed::Box;
