+++ lib/opte/src/engine/predicate.rs
@@ -12,7 +12,6 @@ use super::ether::EtherType;
 use super::ether::EthernetRef;
 use super::icmp::v4::MessageType as IcmpMessageType;
 use super::icmp::v6::MessageType as Icmpv6MessageType;
-use super::ingot_packet::MblkPacketData;
 use super::ip::v4::Ipv4Addr;
 use super::ip::v4::Ipv4Cidr;
 use super::ip::v4::Ipv4Ref;
@@ -21,6 +20,7 @@ use super::ip::v6::Ipv6Addr;
 use super::ip::v6::Ipv6Cidr;
 use super::ip::v6::Ipv6Ref;
 use super::ip::L3;
+use super::packet::MblkPacketData;
 use super::port::meta::ActionMeta;
 use alloc::boxed::Box;
 use alloc::string::String;
