+++ lib/opte/src/engine/snat.rs
@@ -11,9 +11,9 @@ use super::headers::IpMod;
 use super::headers::UlpGenericModify;
 use super::headers::UlpHeaderAction;
 use super::headers::UlpMetaModify;
-use super::ingot_packet::MblkFullParsed;
-use super::ingot_packet::Packet;
 use super::packet::InnerFlowId;
+use super::packet::MblkFullParsed;
+use super::packet::Packet;
 use super::port::meta::ActionMeta;
 use super::predicate::DataPredicate;
 use super::predicate::Predicate;
