+++ lib/oxide-vpc/src/engine/gateway/mod.rs
@@ -56,11 +56,11 @@ use opte::api::Direction;
 use opte::api::OpteError;
 use opte::engine::ether::EtherMod;
 use opte::engine::headers::HeaderAction;
-use opte::engine::ingot_packet::MblkPacketData;
 use opte::engine::layer::DefaultAction;
 use opte::engine::layer::Layer;
 use opte::engine::layer::LayerActions;
 use opte::engine::packet::InnerFlowId;
+use opte::engine::packet::MblkPacketData;
 use opte::engine::port::meta::ActionMeta;
 use opte::engine::port::PortBuilder;
 use opte::engine::port::Pos;
