+++ lib/oxide-vpc/src/engine/mod.rs
@@ -20,10 +20,10 @@ use opte::engine::arp::ValidArpEthIpv4;
 use opte::engine::arp::ARP_HTYPE_ETHERNET;
 use opte::engine::ether::EthernetRef;
 use opte::engine::flow_table::FlowTable;
-use opte::engine::ingot_packet::FullParsed;
-use opte::engine::ingot_packet::Packet;
 use opte::engine::ip::v4::Ipv4Addr;
+use opte::engine::packet::FullParsed;
 use opte::engine::packet::InnerFlowId;
+use opte::engine::packet::Packet;
 use opte::engine::packet::ParseError;
 use opte::engine::parse::ValidGeneveOverV6;
 use opte::engine::parse::ValidNoEncap;
