+++ lib/oxide-vpc/tests/integration_tests.rs
@@ -28,8 +28,6 @@ use opte::engine::ether::Ethernet;
 use opte::engine::ether::EthernetRef;
 use opte::engine::flow_table::FLOW_DEF_EXPIRE_SECS;
 use opte::engine::geneve::Vni;
-use opte::engine::ingot_packet::MblkFullParsed;
-use opte::engine::ingot_packet::Packet;
 use opte::engine::ip::v4::Ipv4Addr;
 use opte::engine::ip::v4::Ipv4Ref;
 use opte::engine::ip::v6::Ipv6;
@@ -37,7 +35,9 @@ use opte::engine::ip::v6::Ipv6Ref;
 use opte::engine::ip::ValidL3;
 use opte::engine::ip::L3;
 use opte::engine::packet::InnerFlowId;
+use opte::engine::packet::MblkFullParsed;
 use opte::engine::packet::MismatchError;
+use opte::engine::packet::Packet;
 use opte::engine::parse::ValidUlp;
 use opte::engine::port::ProcessError;
 use opte::engine::tcp::TcpState;
