+++ xde/src/xde.rs
@@ -64,10 +64,10 @@ use opte::ddi::time::Periodic;
 use opte::engine::ether::EthernetRef;
 use opte::engine::geneve::Vni;
 use opte::engine::headers::IpAddr;
-use opte::engine::ingot_packet::Packet;
 use opte::engine::ioctl::{self as api};
 use opte::engine::ip::v6::Ipv6Addr;
 use opte::engine::packet::InnerFlowId;
+use opte::engine::packet::Packet;
 use opte::engine::packet::ParseError;
 use opte::engine::port::Port;
 use opte::engine::port::PortBuilder;
